FE-62 ✨fix:  주석 수정 및 타입 수정 (#27)
